+++ src/Reconnections/Reconnections.jl
@@ -115,21 +115,30 @@ function reconnect!(
     @timeit to "find reconnection pairs" begin
         to_reconnect = find_reconnection_pairs!(cache, fs, vs; to)
     end
+    isempty(to_reconnect) && return ret_base
+    Fil = eltype(fs)
+    invalidated_filaments = Set{Fil}()  # list of invalidated filaments (can no longer be reconnected)
     @timeit to "reconnect pairs" for reconnect_info ∈ to_reconnect
-        reconnect_info === nothing && continue   # pair has been invalidated
         (; candidate, info,) = reconnect_info
         (; a, b,) = candidate
         @timeit to "reconnect" if a.f === b.f
+            if a.f ∈ invalidated_filaments
+                continue  # don't reconnect this filament if it was already reconnected earlier
+            end
             # Reconnect filament with itself => split filament into two
             @assert a.i ≠ b.i
             nremoved, Lremoved = reconnect_with_itself!(callback, fs, a.f, a.i, b.i, info)
             filaments_removed_count += nremoved
             filaments_removed_length += Lremoved
-            invalidate_segment_pairs!(cache, a.f)
+            push!(invalidated_filaments, a.f)
         else
             # Reconnect two different filaments => merge them into one
+            if a.f ∈ invalidated_filaments || b.f ∈ invalidated_filaments
+                continue  # don't reconnect if one of these filaments was already reconnected earlier
+            end
             reconnect_with_other!(callback, fs, a.f, b.f, a.i, b.i, info)
-            invalidate_segment_pairs!(cache, a.f, b.f)
+            push!(invalidated_filaments, a.f)
+            push!(invalidated_filaments, b.f)
         end
         reconnection_length_loss += info.length_before - info.length_after
         reconnection_count += 1