+++ CHANGELOG.md
@@ -5,6 +5,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add option to inject filaments during a simulation. This should be done by
+  calling the new `Timestepping.inject_filament!` function from within an `affect!` callback
+  (see docs for details).
+
+### Changed
+
+- Reduce variability of results due to recent parallelisation of reconnections.
+  Previously, the order of reconnections was somewhat arbitrary, depending on
+  which threads tested which segment pairs and in which order. We now give
+  priority to pairs which have the smallest distance.
+  This means that results might slightly change compared to simulations with
+  previous versions, since the priority given to reconnection candidates has
+  changed (things are more deterministic now).
+
 ## [0.25.4] - 2024-12-19
 
 ### Changed